# feat: run the optimizer natively