--- a/packages/optimizer/cli/Cargo.toml
+++ b/packages/optimizer/cli/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "penrose-optimizer-cli"
+version = "1.3.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+byte-slice-cast = "1"
+penrose-optimizer = { path = "../core" }
+serde_json = "1"
+wasmer = "3"