--- a/packages/optimizer/core/Cargo.toml
+++ b/packages/optimizer/core/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "penrose-optimizer"
+version = "1.3.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+log = "0.4"
+nalgebra = "0.31"
+serde = { version = "1", features = ["derive"] }
+ts-rs = "6"