--- a/packages/optimizer/instance.d.ts
+++ b/packages/optimizer/instance.d.ts
@@ -1,4 +1,4 @@
-import { InitOutput } from "./build/penrose_optimizer";
+import { InitOutput } from "./build/penrose_optimizer_web";
 
 export declare let maybeOptimizer: InitOutput | undefined;