--- a/packages/optimizer/package.json
+++ b/packages/optimizer/package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "build": "esbuild ./source.ts --outfile=index.js --platform=neutral --bundle --loader:.wasm=binary --define:import.meta.url=null --sourcemap",
     "build-decls": "cargo test",
-    "build-wasm": "cargo build --target=wasm32-unknown-unknown --release",
-    "build-wasm-bindgen": "wasm-bindgen --target=web --keep-lld-exports --out-dir=build target/wasm32-unknown-unknown/release/penrose_optimizer.wasm",
+    "build-wasm": "cargo build --package=penrose-optimizer-web --target=wasm32-unknown-unknown --release",
+    "build-wasm-bindgen": "wasm-bindgen --target=web --keep-lld-exports --out-dir=build target/wasm32-unknown-unknown/release/penrose_optimizer_web.wasm",
     "format": "cargo fmt",
     "format:check": "cargo fmt --check",
     "typecheck": ":"