--- a/packages/optimizer/web/Cargo.toml
+++ b/packages/optimizer/web/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "penrose-optimizer-web"
+version = "1.3.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+console_error_panic_hook = "0.1"
+console_log = "0.2"
+log = "0.4"
+penrose-optimizer = { path = "../core" }
+serde = "1"
+serde-wasm-bindgen = "0.4"
+wasm-bindgen = "0.2"