# Description

Because #1092 rewrote the optimizer in Rust, we now have the ability to compile it not just to WebAssembly but also to native code. This PR provides a CLI for a native Penrose optimizer, which runs a few times faster than the Wasm optimizer.

# Implementation strategy and design decisions

This PR splits the optimizer into three crates:

- `penrose_optimizer` exposes types (with TypeScript declarations generated via ts-rs), constants, and API functions for the optimizer itself, as well as our three autodiff builtins with nontrivial custom implementations (`inverse`, `sign`, and `poly_roots`); the `Compiled` function pointer in the `FnCached` struct is replaced by a `dyn` reference to a `trait` with a `call` function that takes references instead of pointers, which can then be implemented in different ways for different targets
- `penrose_optimizer_web` depends on `penrose_optimizer` and implements the `Compiled` trait for a `struct FuncPtr` that just holds a function pointer like before, then exposes all our Wasm API functions and builtins
- `penrose_optimizer_cli` also depends on `penrose_optimizer`, and uses [Wasmer](https://github.com/wasmerio/wasmer) to compile and run the gradient function, which it reads from a binary Wasm file on disk along with the initial `OptState`

To get those inputs for the CLI, this PR also modifies `core` to put the raw Wasm bytes in the `Gradient`, which `automator` then writes into the output directory along with the initial and optimized diagram states.

# Examples with steps to reproduce them

## Check

To compare the optimized state generated by Wasm vs native, run these commands:

```sh
yarn
yarn build:automator
cd packages/automator
yarn start batch registry.json out/ --src-prefix=../examples/src/ --folders
cd ../optimizer
cargo run --release ../automator/out/tree-venn{,.json}
git diff --no-index ../automator/out/tree-venn{/optimized,}.json
```

## Speedup

To see optimization times for all our registry diagrams, run these commands:

```sh
yarn
yarn build:automator
cd packages/automator
yarn start batch registry.json out/ --src-prefix=../examples/src/ --folders
cd ../optimizer
cargo build --release
for example in ../automator/out/*; do
  echo $example
  time target/release/penrose-optimizer-cli $example
  echo
done
```

Output:

```
../automator/out/3d-projection-fake-3d-linear-algebra
target/release/penrose-optimizer-cli $example  0.04s user 0.01s system 199% cpu 0.023 total

../automator/out/aggregateData.json
thread 'main' panicked at 'called `Result::unwrap()` on an `Err` value: Os { code: 20, kind: NotADirectory, message: "Not a directory" }', cli/src/main.rs:80:10
note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
target/release/penrose-optimizer-cli $example  0.00s user 0.00s system 72% cpu 0.004 total

../automator/out/allShapes-allShapes
target/release/penrose-optimizer-cli $example  0.04s user 0.01s system 348% cpu 0.013 total

../automator/out/arrowheads-arrowheads
target/release/penrose-optimizer-cli $example  0.03s user 0.01s system 333% cpu 0.010 total

../automator/out/circle-example-euclidean
target/release/penrose-optimizer-cli $example  0.26s user 0.01s system 133% cpu 0.198 total

../automator/out/collinear-euclidean
target/release/penrose-optimizer-cli $example  0.05s user 0.00s system 204% cpu 0.026 total

../automator/out/congruent-triangles-euclidean
target/release/penrose-optimizer-cli $example  0.12s user 0.01s system 491% cpu 0.025 total

../automator/out/continuousmap-continuousmap
target/release/penrose-optimizer-cli $example  0.03s user 0.00s system 342% cpu 0.010 total

../automator/out/hypergraph-hypergraph
target/release/penrose-optimizer-cli $example  1.37s user 0.51s system 110% cpu 1.702 total

../automator/out/incenter-triangle-euclidean
target/release/penrose-optimizer-cli $example  0.10s user 0.01s system 215% cpu 0.047 total

../automator/out/lagrange-bases-lagrange-bases
target/release/penrose-optimizer-cli $example  0.03s user 0.00s system 323% cpu 0.010 total

../automator/out/midsegment-triangles-euclidean
target/release/penrose-optimizer-cli $example  0.05s user 0.00s system 440% cpu 0.012 total

../automator/out/non-convex-non-convex
target/release/penrose-optimizer-cli $example  0.14s user 0.01s system 291% cpu 0.051 total

../automator/out/one-water-molecule-atoms-and-bonds
target/release/penrose-optimizer-cli $example  0.01s user 0.00s system 253% cpu 0.006 total

../automator/out/parallel-lines-euclidean
target/release/penrose-optimizer-cli $example  0.05s user 0.01s system 228% cpu 0.026 total

../automator/out/persistent-homology-persistent-homology
target/release/penrose-optimizer-cli $example  1.31s user 0.49s system 106% cpu 1.696 total

../automator/out/points-around-line-shape-distance
target/release/penrose-optimizer-cli $example  0.07s user 0.01s system 343% cpu 0.021 total

../automator/out/points-around-polyline-shape-distance
target/release/penrose-optimizer-cli $example  0.14s user 0.01s system 483% cpu 0.031 total

../automator/out/points-around-star-shape-distance
target/release/penrose-optimizer-cli $example  0.24s user 0.01s system 562% cpu 0.045 total

../automator/out/siggraph-teaser-euclidean-teaser
target/release/penrose-optimizer-cli $example  0.07s user 0.01s system 353% cpu 0.022 total

../automator/out/small-graph-disjoint-rect-line-horiz
target/release/penrose-optimizer-cli $example  0.21s user 0.01s system 575% cpu 0.039 total

../automator/out/small-graph-disjoint-rects
target/release/penrose-optimizer-cli $example  0.02s user 0.00s system 377% cpu 0.008 total

../automator/out/small-graph-disjoint-rects-large-canvas
target/release/penrose-optimizer-cli $example  0.02s user 0.01s system 406% cpu 0.007 total

../automator/out/small-graph-disjoint-rects-small-canvas
target/release/penrose-optimizer-cli $example  0.02s user 0.00s system 365% cpu 0.008 total

../automator/out/tree-tree
target/release/penrose-optimizer-cli $example  0.03s user 0.00s system 393% cpu 0.009 total

../automator/out/tree-venn
target/release/penrose-optimizer-cli $example  0.04s user 0.00s system 314% cpu 0.015 total

../automator/out/tree-venn-3d
target/release/penrose-optimizer-cli $example  0.06s user 0.01s system 319% cpu 0.021 total

../automator/out/two-vectors-perp-vectors-dashed
target/release/penrose-optimizer-cli $example  0.02s user 0.01s system 323% cpu 0.009 total

../automator/out/vector-wedge-exterior-algebra
target/release/penrose-optimizer-cli $example  0.03s user 0.01s system 365% cpu 0.011 total

../automator/out/wet-floor-atoms-and-bonds
target/release/penrose-optimizer-cli $example  2.69s user 0.01s system 101% cpu 2.663 total

../automator/out/word-cloud-example-word-cloud
target/release/penrose-optimizer-cli $example  0.11s user 0.00s system 303% cpu 0.037 total

../automator/out/wos-laplace-estimator-walk-on-spheres
target/release/penrose-optimizer-cli $example  0.17s user 0.01s system 429% cpu 0.041 total

../automator/out/wos-nested-estimator-walk-on-spheres
target/release/penrose-optimizer-cli $example  0.72s user 0.01s system 163% cpu 0.445 total

../automator/out/wos-offcenter-estimator-walk-on-spheres
target/release/penrose-optimizer-cli $example  0.21s user 0.01s system 405% cpu 0.053 total

../automator/out/wos-poisson-estimator-walk-on-spheres
target/release/penrose-optimizer-cli $example  0.31s user 0.01s system 223% cpu 0.142 total
```

## Flamegraph

We can also use this native optimizer can be used to generate flamegraphs. [Install `cargo-flamegraph`](https://github.com/flamegraph-rs/flamegraph/tree/b65fd831c825d167ceda0c94055bba782b27da49#installation), then run these commands (note, the `--root` option is only necessary if you're on macOS):

```sh
yarn
yarn build:automator
cd packages/automator
yarn start batch registry.json out/ --src-prefix=../examples/src/ --folders
cd ../optimizer
cargo flamegraph --root -- ../automator/out/persistent-homology-persistent-homology
```

Output (in `packages/optimizer/flamegraph.svg`):

![flamegraph](https://user-images.githubusercontent.com/8246041/211158584-f622daa5-8726-4b2a-bf15-0092f5ef18fa.svg)

# Checklist

- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new ESLint warnings
- [x] I have reviewed any generated registry diagram changes