@samestep 2023-07-13
@keenancrane No this doesn't have anything to do with Rose; just closing because it didn't seem like there was any interest. We can always reopen this if we want to.